backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
beautifulsoup4==4.12.2
google-generativeai==0.3.1
python-dotenv==1.0.0
google-api-python-client==2.108.0
python-socketio==5.8.0
flask-socketio==5.3.6 